+++ tests/testthat/test-sample_things.R
@@ -1,6 +1,6 @@
 test_that("covariance matrix can be sampled", {
   expect_silent(
-    assert_covariance_matrix(sample_covariance_matrix(10))
+    assert_covariance_matrix(sample_covariance_matrix(3))
   )
   expect_silent(
     assert_covariance_matrix(sample_covariance_matrix(3, diag = TRUE))