+++ tests/testthat/test-timed_evaluation.R
@@ -5,7 +5,7 @@ test_that("interruption of long evaluations works", {
   }
   expect_error(
     timed(foo(0.5), "1"),
-    "Assertion on 'secs' failed: Must be of type 'number', not 'character'."
+    "Assertion on 'seconds' failed: Must be of type 'number', not 'character'."
   )
   expect_equal(timed(foo(0.5), 5), 0.5)
   expect_null(timed(foo(5), 1))